+++ components/mail-form/mail-form.tsx
@@ -14,23 +14,11 @@ import { useForm } from "react-hook-form";
 import { zodResolver } from "@hookform/resolvers/zod";
 import { formSchema } from "@/lib/formSchema";
 import { Textarea } from "../ui/textarea";
+import useMailForm from "@/hooks/useMailForm";
 
 export default function MailForm() {
-  const form = useForm({
-    resolver: zodResolver(formSchema),
-    defaultValues: {
-      username: "",
-      subject: "",
-      email: "",
-      content: "",
-    },
-  });
 
-  function onSubmit(values: any) {
-    // Do something with the form values.
-    // ✅ This will be type-safe and validated.
-    console.log(values);
-  }
+  const {form, onSubmit} = useMailForm()
 
   return (
     <Form {...form}>
